Change the way the talker process is terminated.
Instead of killing it, the fact that it has to stop sending messages is
signaled through the `State`. This will prevent the `talker` process from being
killed without sending the messages to all the neighbors. Note that
inconsistencies in the number of messages might still arise if the grace period
is too short.